requirements.txt
pillow
numpy

pytest
